Authorization user CRUD

+++ src/tasks/tasks.service.ts
@@ -42,14 +42,25 @@ export class TasksService {
     return this.taskRepository.createTask(createTaskDto, user);
   }
 
-  // async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-  //   const task = await this.getTaskByID(id);
-  //   task.status = status;
-  //   await task.save();
-  //   return task;
-  // }
+  async updateTaskStatus(
+    id: number,
+    status: TaskStatus,
+    user: User,
+  ): Promise<Task> {
+    const task = await this.getTaskByID(id, user);
+    task.status = status;
+    await task.save();
+    return task;
+  }
 
-  async deleteTask(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTask(
+    id: number,
+    user: User,
+  ): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
+
+    if (result.affected === 0 ) {
+      throw new NotFoundException(`Task with ID "${id}" not found`);
+    }
   }
 }
